+++ CHANGES.rst
@@ -5,6 +5,7 @@ Version 0.9.0 (UNRELEASED)
 --------------------------
 
 - Adds new ``/api/launch`` endpoint that allows running workflows from remote sources.
+- Adds the possibility to fetch the workflow specification from a remote URL.
 
 Version 0.8.4 (2022-02-23)
 --------------------------